pyproject.toml
[project]
name = "application-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=1.7.1",
    "langchain>=1.2.12",
    "langchain-huggingface>=1.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "transformers>=5.3.0",
]
